requirements.txt
streamlit>=1.29.0
mysql-connector-python>=8.1.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
pillow>=10.0.0
streamlit-option-menu>=0.3.6
